refactor(netsync): Replication.cpp becomes seven files, one subject each
3,688 lines in one translation unit, covering both halves of the protocol and
everything from config loading to interpolation. Nothing here changes what the
code does — every line moves, none is edited — but the seams the file already
had in comment banners are now real boundaries.

  ReplicationConfig.cpp          the relevancy config loader
  ReplicationServer.cpp          lifecycle, connections, control, ids, relevancy
  ReplicationServerMessages.cpp  inbound messages, input, intents, events
  ReplicationServerState.cpp     the per-tick bookkeeping behind a snapshot
  ReplicationServerSnapshot.cpp  writing and sending a snapshot
  ReplicationClient.cpp          lifecycle, messages, applying a snapshot
  ReplicationClientMotion.cpp    mirror bodies, smoothing, interpolation

The four key-packing helpers were an anonymous-namespace block shared by both
halves, so they move to src/ReplicationInternal.hpp as inline functions rather
than becoming a static copy per TU. EnsureComponent has one caller, in
ApplySnapshot, so it travels with the client. IntentGate is used only by
HandleIntent and stays anonymous in the messages TU.

83 member-function definitions before, the same 83 after. Build clean, ctest
14/14.
